+++ app/src/main/java/io/github/sds100/keymapper/mappings/keymaps/trigger/BaseConfigTriggerViewModel.kt
@@ -754,7 +754,9 @@ abstract class BaseConfigTriggerViewModel(
     }
 
     fun onEnableGuiKeyboardClick() {
-        setupGuiKeyboard.enableInputMethod()
+        coroutineScope.launch {
+            setupGuiKeyboard.enableInputMethod()
+        }
     }
 
     fun onChooseGuiKeyboardClick() {