+++ app/src/main/java/io/github/sds100/keymapper/system/accessibility/IAccessibilityService.kt
@@ -49,6 +49,8 @@ interface IAccessibilityService {
     val rootNode: AccessibilityNodeModel?
     val activeWindowPackage: Flow<String?>
 
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
+    fun setInputMethodEnabled(imeId: String, enabled: Boolean)
     fun hideKeyboard()
     fun showKeyboard()
     val isKeyboardHidden: Flow<Boolean>